requirements.txt
asgiref==3.3.1
Django==3.1.4
django-enumfield==2.0.2
pytz==2020.5
six==1.15.0
sqlparse==0.4.1

django-crispy-forms~=1.11.1